+++ app/controllers/v1/SubjectController.java
@@ -10,7 +10,7 @@ public class SubjectController extends BaseController {
     @Inject
     private SubjectDao subjectDao;
 
-    public Result show(Integer id) {
+    public Result show(Long id) {
         Subject s = subjectDao.find(id);
 
         return ok(gson.toJson(s));
